+++ ART_TELEMETRY_APP/ART_TELEMETRY_APP/Charts/Classes/OrderBy.cs
@@ -5,7 +5,7 @@ namespace ART_TELEMETRY_APP.Charts.Classes
 {
     struct OrderBy
     {
-        public List<float> Data;
+        public List<double> Data;
         public int FromIndex;
         public int ToIndex;
     }
